Setup new test class and tests
